% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_bbref.R
\name{scrape_bbref}
\alias{scrape_bbref}
\title{Script to scrape per 100 possession data from basketball reference}
\usage{
scrape_bbref(season)
}
\arguments{
\item{season}{A numeric object that defines the ending year of the NBA season}
}
\value{
A tibble of per 100 possession stats from Basketball Reference for NBA players in the given season
}
\description{
Script to scrape per 100 possession data from basketball reference
}
